third_party/blink/web_tests/fast/dom/htmlcollection-reachable-expected.txt

PASS document.anchors.customProperty0 is 0
PASS document.applets.customProperty1 is 1
PASS document.embeds.customProperty2 is 2
PASS document.forms.customProperty3 is 3
PASS document.images.customProperty4 is 4
PASS document.links.customProperty5 is 5
PASS document.plugins.customProperty6 is 6
PASS document.scripts.customProperty7 is 7
PASS element.children.customProperty8 is 8
PASS fieldset.elements.customProperty9 is 9
PASS form.elements.customProperty10 is 10
PASS map.areas.customProperty11 is 11
PASS section.rows.customProperty12 is 12
PASS select.selectedOptions.customProperty13 is 13
PASS table.rows.customProperty14 is 14
PASS table.tBodies.customProperty15 is 15
PASS tableRow.cells.customProperty16 is 16
PASS successfullyParsed is true

TEST COMPLETE
